requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0.1
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.12.1
django-advanced-filters==2.0.0
django-cors-headers==4.3.1
django-crum==0.7.9
django-editorjs==0.2.1
django-environ==0.11.2
django-filter==23.5
django-import-export==3.3.7
django-jazzmin==2.6.1
django-object-actions==4.2.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
iniconfig==2.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
pytz==2024.1
PyYAML==6.0.1
simplejson==3.19.2
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
